% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.PRELES.R
\name{write.config.PRELES}
\alias{write.config.PRELES}
\title{Write PRELES configuration files}
\usage{
write.config.PRELES(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
configuration file for PRELES for given run
}
\description{
Writes a PRELES config file.
}
\author{
Tony Gardella, Micheal Dietze
}
