% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPRELES.jobsh.R
\name{runPRELES.jobsh}
\alias{runPRELES.jobsh}
\title{Process ncdf file, run PRELES model, and convert output .nc file in CF standard}
\usage{
runPRELES.jobsh(
  met.file,
  outdir,
  parameters,
  sitelat,
  sitelon,
  start.date,
  end.date
)
}
\arguments{
\item{met.file}{base name for yearly nc files containing met data.
Example: `met.file="somefile"` matches somefile.2004.nc, somefile.2005.nc, etc.}

\item{outdir}{Location of PRELES model output}

\item{parameters}{An R data file containing parameter values.
Must be an Rda file written via `save()`, and must define an object named
`trait.values`}

\item{sitelat, sitelon}{Latitude and longitude of site in decimal degrees}

\item{start.date, end.date}{Start and end time of the simulation}
}
\description{
Process ncdf file, run PRELES model, and convert output .nc file in CF standard
}
\author{
Tony Gardella, Michael Dietze
}
